- corrected Via handling for transactional replies to correspond to RFC3261, section 18.2.2. - adds support for 'received' (added on transactional replies) - adds support for 'rport' (RFC 3581)
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@1713 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ core/plug-in/sipctrl/hash_table.cpp
@@ -124,7 +124,7 @@ sip_trans* trans_bucket::match_request(sip_msg* msg)
 		continue;
 	    }
 	    
-	    if((*it)->msg->via_p1->branch.len != len + MAGIC_BRANCH_LEN)
+	    if((*it)->msg->via_p1->branch.len != (unsigned int)len + MAGIC_BRANCH_LEN)
 		continue;
 
 	    if((*it)->msg->via_p1->host.len !=
